Change AttestationPropagationSlotRange type to Integer

+++ data/provider/src/main/java/tech/pegasys/teku/api/GetSpecResponse.java
@@ -119,8 +119,7 @@ public class GetSpecResponse {
   }
 
   private String getAttestationPropagationSlotRange() {
-    return Integer.toString(
-        specConfig.getNetworkingConfig().getAttestationPropagationSlotRange().intValue());
+    return Integer.toString(specConfig.getNetworkingConfig().getAttestationPropagationSlotRange());
   }
 
   private String getMaximumGossipClockDisparity() {
