+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/NetworkingSpecConfig.java
@@ -14,7 +14,6 @@
 package tech.pegasys.teku.spec.config;
 
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
 public interface NetworkingSpecConfig {
 
@@ -34,7 +33,7 @@ public interface NetworkingSpecConfig {
   // in seconds
   int getRespTimeout();
 
-  UInt64 getAttestationPropagationSlotRange();
+  int getAttestationPropagationSlotRange();
 
   // in millis
   int getMaximumGossipClockDisparity();
